在庫管理（NETSEA用）/requirements.txt
black==24.10.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.15
Flask==3.1.0
Flask-Limiter==3.8.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
schedule==1.2.2
setuptools==75.6.0
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.0